+++ CHANGELOG.md
@@ -94,6 +94,7 @@
 - Fixed content exclusion when extracting DMG archives of IntelliJ Platform on macOS
 - Could not find a field for name `metadata/modelVersion` (Attribute) in `MavenMetadata`
 - `PluginArtifactoryShim`: use only host when setting up proxy for custom plugin repositories
+- Fixed searchable options resolving on `2024.2+`
 
 ## [2.0.0-beta6] - 2024-06-06